Everything now builds with VC++2008
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@8635 7a7537e8-13f0-0310-91df-b6672ffda945

+++ cups/dir.c
@@ -73,7 +73,7 @@ _cups_dir_time(FILETIME ft)		/* I - File time */
   * between them...
   */
 
-  val = ft.dwLowDateTime + (ft.dwHighDateTime << 32);
+  val = ft.dwLowDateTime + ((ULONGLONG)ft.dwHighDateTime << 32);
   return ((time_t)(val / 10000000 - 11644732800));
 }
